requirements.txt
boto3==1.37.0
botocore==1.37.0
fastapi==0.95.2
huggingface_hub==0.25.2
pydantic==1.10.9
python-dotenv==1.0.1
sentence_transformers==2.2.2
tensorflow==2.13.0
uvicorn==0.34.0
typing-extensions==4.5.0